requirements.txt
certifi==2025.8.3
charset-normalizer==3.4.3
idna==3.10
psutil==7.0.0
PySocks==1.7.1
python-dotenv==1.1.1
requests==2.32.5
stem==1.8.2
urllib3==2.5.0
